# elm-frontend/docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "吃饱了吗点餐平台",
    "description": "Tianjin University Software Engineering Project",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/reviews/order/{orderId}": {
      "post": {
        "summary": "添加评价",
        "deprecated": false,
        "description": "用户评价订单",
        "tags": ["管理评价"],
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "integer",
                    "format": "int64"
                  },
                  "createTime": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "updateTime": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "creator": {
                    "type": "integer",
                    "format": "int64"
                  },
                  "updater": {
                    "type": "integer",
                    "format": "int64"
                  },
                  "deleted": {
                    "type": "boolean"
                  },
                  "customer": {
                    "$ref": "#/components/schemas/User"
                  },
                  "business": {
                    "$ref": "#/components/schemas/Business"
                  },
                  "order": {
                    "$ref": "#/components/schemas/Order"
                  },
                  "anonymous": {
                    "type": "boolean"
                  },
                  "stars": {
                    "type": "integer",
                    "format": "int32"
                  },
                  "content": {
                    "type": "string"
                  }
                },
                "required": ["stars", "id", "content", "anonymous"]
              },
              "examples": {}
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "code": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Review"
                    },
                    "message": {
                      "type": "string",
                      "nullable": true
                    }
                  },
                  "required": ["data"]
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "get": {
        "summary": "获取订单评价 ",
        "deprecated": false,
        "description": "获取指定订单评价",
        "tags": ["管理评价"],
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "null"
              },
              "examples": {}
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "code": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Review"
                    },
                    "message": {
                      "type": "string",
                      "nullable": true
                    }
                  },
                  "required": ["data"]
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/reviews/{reviewId}": {
      "patch": {
        "summary": "修改评价 ",
        "deprecated": false,
        "description": "用户修改评价（仅能修改stars, content和anonymous）",
        "tags": ["管理评价"],
        "parameters": [
          {
            "name": "reviewId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "integer",
                    "format": "int64"
                  },
                  "createTime": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "updateTime": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "creator": {
                    "type": "integer",
                    "format": "int64"
                  },
                  "updater": {
                    "type": "integer",
                    "format": "int64"
                  },
                  "deleted": {
                    "type": "boolean"
                  },
                  "customer": {
                    "$ref": "#/components/schemas/User"
                  },
                  "business": {
                    "$ref": "#/components/schemas/Business"
                  },
                  "order": {
                    "$ref": "#/components/schemas/Order"
                  },
                  "anonymous": {
                    "type": "boolean"
                  },
                  "stars": {
                    "type": "integer",
                    "format": "int32"
                  },
                  "content": {
                    "type": "string"
                  }
                }
              },
              "examples": {}
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "code": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Review"
                    },
                    "message": {
                      "type": "string",
                      "nullable": true
                    }
                  },
                  "required": ["data"]
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "delete": {
        "summary": "删除评价",
        "deprecated": false,
        "description": "用户或管理员删除评价",
        "tags": ["管理评价"],
        "parameters": [
          {
            "name": "reviewId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "null"
              },
              "examples": {}
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "code": {
                      "type": "string"
                    },
                    "data": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string",
                      "nullable": true
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/api/reviews/business/{businessId}": {
      "get": {
        "summary": "获取店铺全部订单评价",
        "deprecated": false,
        "description": "获取店铺全部订单评价，若用户设置匿名，则order, customer, creator置null",
        "tags": ["管理评价"],
        "parameters": [
          {
            "name": "businessId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "null"
              },
              "examples": {}
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "code": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Review"
                      }
                    },
                    "message": {
                      "type": "string",
                      "nullable": true
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    }
  },
  "components": {
    "schemas": {
...
      "Authority": {
        "required": ["name"],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
      "Business": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "createTime": {
            "type": "string",
            "format": "date-time"
          },
          "updateTime": {
            "type": "string",
            "format": "date-time"
          },
          "creator": {
            "type": "integer",
            "format": "int64"
          },
          "updater": {
            "type": "integer",
            "format": "int64"
          },
          "deleted": {
            "type": "boolean"
          },
          "businessName": {
            "type": "string"
          },
          "businessOwner": {
            "$ref": "#/components/schemas/User"
          },
          "businessAddress": {
            "type": "string"
          },
          "businessExplain": {
            "type": "string"
          },
          "businessImg": {
            "type": "string"
          },
          "orderTypeId": {
            "type": "integer",
            "format": "int32"
          },
          "startPrice": {
            "type": "number"
          },
          "deliveryPrice": {
            "type": "number"
          },
          "remarks": {
            "type": "string"
          }
        },
        "required": ["id"]
      },
      "Review": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "createTime": {
            "type": "string",
            "format": "date-time"
          },
          "updateTime": {
            "type": "string",
            "format": "date-time"
          },
          "creator": {
            "type": "integer",
            "format": "int64"
          },
          "updater": {
            "type": "integer",
            "format": "int64"
          },
          "deleted": {
            "type": "boolean"
          },
          "customer": {
            "$ref": "#/components/schemas/User"
          },
          "business": {
            "$ref": "#/components/schemas/Business"
          },
          "order": {
            "$ref": "#/components/schemas/Order"
          },
          "anonymous": {
            "type": "boolean"
          },
          "stars": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "string"
          }
        },
        "required": ["stars", "id"]
      },
      "User": {
        "required": ["username"],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "createTime": {
            "type": "string",
            "format": "date-time"
          },
          "updateTime": {
            "type": "string",
            "format": "date-time"
          },
          "creator": {
            "type": "integer",
            "format": "int64"
          },
          "updater": {
            "type": "integer",
            "format": "int64"
          },
          "deleted": {
            "type": "boolean"
          },
          "username": {
            "maxLength": 100,
            "minLength": 1,
            "type": "string"
          },
          "authorities": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Authority"
            }
          }
        }
      },
...
      "DeliveryAddress": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "createTime": {
            "type": "string",
            "format": "date-time"
          },
          "updateTime": {
            "type": "string",
            "format": "date-time"
          },
          "creator": {
            "type": "integer",
            "format": "int64"
          },
          "updater": {
            "type": "integer",
            "format": "int64"
          },
          "deleted": {
            "type": "boolean"
          },
          "contactName": {
            "type": "string"
          },
          "contactSex": {
            "type": "integer",
            "format": "int32"
          },
          "contactTel": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "customer": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
      "Order": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "createTime": {
            "type": "string",
            "format": "date-time"
          },
          "updateTime": {
            "type": "string",
            "format": "date-time"
          },
          "creator": {
            "type": "integer",
            "format": "int64"
          },
          "updater": {
            "type": "integer",
            "format": "int64"
          },
          "deleted": {
            "type": "boolean"
          },
          "customer": {
            "$ref": "#/components/schemas/User"
          },
          "business": {
            "$ref": "#/components/schemas/Business"
          },
          "orderDate": {
            "type": "string",
            "format": "date-time"
          },
          "orderTotal": {
            "type": "number"
          },
          "deliveryAddress": {
            "$ref": "#/components/schemas/DeliveryAddress"
          },
          "orderState": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
    },
    "securitySchemes": {
      "JWT Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "servers": [],
  "security": []
}